Revert "Remove impl header from public header"
This reverts commit 9428

This commit breaks build on windows because HANDLE is not defined.
Including WinNT.h or Windows.h does not work, including ace/config.h
works, but in my opinion a public header should not include a file
from ace (ace headers should not be required to build a program using
just Yarp), so I'm reverting this commit until I come out with a
better solution.

svn path=/trunk/yarp2/; revision=9430