Do not use this in initialization list
To do this, use pointers instead of references, and make owner public
(the class itself is private so it cannot be accessed from outside
anyway).

As visual studio reports, passing this in the initialization list is
dangerous, because the object pointed by "this" is not created yet.
Therefore if the pointer is used in the member class constructor, the
behaviour is unpredictable.
This should fix one of those msvc warning.

svn path=/branches/yarp-2.4/; revision=9967