[yarprobotinterface] Parse DTD tag from the xml files
Depending on the file type it must be something like:

<!DOCTYPE robot PUBLIC "-//YARP//DTD yarprobotinterface 1.0//EN" "http://www.icub.org/DTD/yarprobotinterfaceV1.0.dtd">
<!DOCTYPE devices PUBLIC "-//YARP//DTD yarprobotinterface 1.0//EN" "http://www.icub.org/DTD/yarprobotinterfaceV1.0.dtd">
<!DOCTYPE params PUBLIC "-//YARP//DTD yarprobotinterface 1.0//EN" "http://www.icub.org/DTD/yarprobotinterfaceV1.0.dtd">

The URL tag is parsed to extract the version number that will be used
to check the version of the xml syntax.
For now the only check is that the main file is 1.0 and that for the
other files the major number is the same of the main file. In future
more complex checks can be added.

svn path=/trunk/iCub/; revision=25972