Propagate writeStrict port option to RemoteCotrolBoard
yarp port has the writeStrict option but was not possible to tell RemoteControlBoard to activate it if necessary.
The configuration now checks also for a 'writeStrict' option with 'on' or 'off' as possible choises where 'off' is the default (keep backward behaviour).
This happened to be needed for streaming port because otherwise some commands were sistematically lost.