Forbid CMake 2.8.11
CMake 2.8.11 fails on some systems with a very obscure error.
This is due to a bug when quoting libraries in try_compile, and it was
fixed in 2.8.11.1. Older versions do not seem to be this problem.
In order to ensure that the user do not use the bugged version (and
consequently complain about not being able to build the project) we
ensure here that the version is different from 2.8.11 and print a
readable error that the user will be able to understand.

See also: http://cmake.org/gitweb?p=cmake.git;a=commit;h=e65ef08bf2719ffd1cc4226f9594ff7127ad8b5e