New wrappers imported from iCub repository
New wrappers are:
-controlBoardWrapper2
-analogWrapper         (dataflow from the robot the the external world)
-virtualAnalogWrapper  (dataflow from the external world to the robot)

This is under testing, the flag 'NEW_WRAPPERS_WIP' has to be set in order to compile this new devices.
Enabling this flag should not break the compilation of iCub repository since both devices can live side by side.

When both are enabled the factory will choose whatever came first, it means that the yarp one will be
created since it'll be always compiled first.