[ros] add --ros option to nameserver.
This option opens a back-channel to a ROS name-server, and
delegates all persistence to it for ports that ROS can name.
Requires the ROS_MASTER_URI environment variable to be set.
The same behavior can be triggered by setting the YARP_USE_ROS
environment variable before starting the YARP name server.
This is just one more of several ways always available to use
ROS with YARP.