Use relative path instead of absolute to install python bindings
Python files are the only ones installed using the absolute path, but
nonetheless they are installed in the prefix.

This change forces CMake to handle the destination for these files as
relative (like for all the other installed files) instead of as
absolute. The install script is therefore relocatable by setting
a different CMAKE_PREFIX_PATH.

This allows to perform a proper installation in /foo instead of the
prefix set during configure phase by running

  cmake CMAKE_PREFIX_PATH=/foo -P cmake_install.cmake

without configuring and building the whole project again.