Fix rendering images with padding.
Padding was not considered when rendering images.

Qt 5.3.0 and previous version are bugged and the image will not be
displayed correctly anyway.

Qt 5.3.1 is partially fixed, but unfortunately it is still bugged.

A patch to fix this bug in Qt5 was already submitted upstream:

   https://codereview.qt-project.org/#/c/92095/

Therefore, (assuming that the patch will be merged in Qt 5.3.2)
depending on Qt5 version, the code does a memcopy of the whole image
(Qt 5.3.2 or later) or a line by line memcopy discarding the padding.