yarp plugins: add generation of .ini manifest file in YARP_PREPARE_DEVICE
The macro YARP_PREPARE_DEVICE now will automatically generate the
ini manifest file from the devices using the information from the
parameter of the macro itself and the template located in
/conf/template/YarpPluginConfFile.cmake.in

The file name will be ${plugin_name}_manifest.ini, and will be located
in the source folder of the plugin. It can be installed later by using
yarp_install macro.

This commit also add an optional parameter in the YARP_PREPARE_DEVICE macro
to store the name of the library (.so file) in case it is different from
the plugin_name, otherwise is it assume to be equal to the plugin name.