[swig] fix translation of Contactable::setEnvelope, fixes #372
This method expects a PortWriter. Users will typically pass a Portable
which is both a PortReader and a PortWriter. Swig has a tough time
translating multiple inheritance (for good reason), so translations
may not know that Portables are PortWriters (see #372 for example).
We can help out by explicitly accepting Portables.