[CMake] Deprecate YARP_ADD_CARRIER_FINGERPRINT and YARP_ADD_DEVICE_FINGERPRINT
Both methods puts the build tree files in the wrong directory (should
use the name passed to YARP_CONFIGURE_EXTERNAL_INSTALLATION) that is
correct only for YARP.

The install command is performed only for YARP, moreover it is not clear
from the macro name that it will copy/install the file, explicitly using
YARP_INSTALL is more clear and the correct destination can be passed to
external installations. Also since the installation is not performed,
another INSTALL or YARP_INSTALL command is already required.

iCub already uses YARP_INSTALL for these reasons.

Finally, the additional arguments are useless and are just discarded and
might cause confusion or be outdated, therefore there isn't any real
advantage in using these macros.