Make class name available for plugin devices
This info is already available in the library, but it is not exposed for
use.
There are probably not many reasons for exposing this information, but
without this, the plugin factory ignores the name of the class for the
plugins, and therefore the documentation generated by harness_dev is
different.