[CMake] Improve plugin installation
Add install dir variables for plugins and explicitly install them

Handle TARGETS argument in YARP_INSTALL. Normal targets are just
installed, plugins targets are fixed so that they are built and
installed in the proper directory, and are not EXPORTed.

Dynamic plugins are now installed in lib/${YARP_PACKAGE_NAME} instead of
just in lib. This makes it easier to understand what is a library and
what is a plugin.

YARP plugins are no longer automatically installed by YarpPlugin.cmake
but manually in the right place.