[YARP_OS] Add methods to send envelope with the stream
This allows envelope to be carried inside the stream instead of by
administrative messages.

The carrier will have to overload the yarp::os::Carrier::handleEnvelope
method (that by default does nothing).
This method will be called by the PortCoreOutputUnit for non-escapable
Carriers (like the mjpeg carrier) instead of sending the administrative
command containing the envelope.

On the receiving side, the PortCoreInputUnit will install a callback
using the yarp::os::InputStream::setReadEnvelopeCallback method (that
should be overloaded by non-escapable carriers that want to handle the
envelope).
The input stream must then execute the callback as soon as the envelope
is ready.
This is necessary, unless we want break compatibility, because a
delegate could be used to read objects from the network (basically we
should modify any class derived by PortReader), and when the control
comes back to the PortCoreInputUnit (i.e. after the readBlock() call) it
is too late to set the envelope.
For example, when the readBlock() returns, "yarp read ... envelope" will
already have displayed a message without the envelope (for the first
message) or using the envelope from the previous messages.