[mjpeg] Send envelope inside the images
The COM special marker is used during jpeg compression for this purpose.
See also libjpeg documentation

  http://svn.code.sf.net/p/libjpeg-turbo/code/trunk/libjpeg.txt

On the receiving side, there is no check on the message received.
This is intentional and it means that if some web camera sends some
comment using the COM marker, the receiver will interpret it as the
envelope.
If people starts complaining about this, perhaps we could choose one of
the APP0 through APP15 marker instead, or perhaps using some sort of
protocol inside the envelope to ensure that we are receiving an envelope
and not some random comment. Anyway for now this shouldn't be a big
issue.