[YARP_OS] Fix MSVC C4717 warning in yarp::os::Value
warning C4717: 'yarp::os::Value::operator==' : recursive on all control
paths, function will cause runtime stack overflow

This warning was introduced by ef3173560105e609a5e31c04574fa8039392b950

Use a yarp::os::impl::Storable instead of another Value as proxy. The
proxy was already guaranteed to be a Storable (i.e. a class derived by
Value), but it was saved as a Value pointer, probably to hide the
existence of the Storable class.
The Storable is nonetheless hidden, since it just needs to be
forward-declared in the class declaration.
At some point it should disappear again, and should be hidden using the
PIMPL idiom.

This change simplifies a lot understanding how the Value class works.