Fix race condition when calling yarp::os::Thread::getKey() method
When calling getKey() in the run() function of a Thread, when the
std::thread is used, the id variable could be uninitialized, since the
thread is started before this variable is set. The same issue happens
when SKIP_ACE is enabled and therefore pthread is used.

The ACE thread implementation does not have this issue, since the id
variable is set automatically by the ACE_Thread::spawn() method.