Fix some cases of connections leaving a streaming connection pending
When connecting two ports, if a connection already exists between these two
ports, this is disconnected and replaced with the new connection.
This usually happens on the same side, for example if a connection is "udp" and
a "tcp" connection is requested, the sending port is contacted, and the
connection is closed.

When connecting using a "pull" carrier (i.e. mjpeg), if a connection between
the two ports exists it is closed, but on the receiver side. This is not a
problem for "connected" connections, like tcp, because the sender notices that
the receiver cut the connection, but for "connectionless" connections (like udp
and mcast), the receiver has no way to know that the receiver stopped listening
and therefore it cannot know that it should stop streaming data. (The same thing
should theoretically happen with a "pull-connectionless" carrier followed
by a normal one)

This means that the port continues streaming data that is not read by anyone
until a disconnect command is sent.

The process executing the disconnection does not necessarily know anything about
the carriers (for example mjpeg could not be built on the system calling "yarp
connect"), therefore since now this information is returned by the PortCore list
[in|out] commands (see e9295b3671f3114a65421c66c53c1f7838de03b7), we use this
information to detect the cases when the disconnection will not happen
automatically, and eventually we send a disconnection command to the port
sending data.

This means that each "connect" and "disconnect" command needs to execute an
extra "list" command, but unfortunately there is no other way to know whether a
disconnect command should be sent or not.