YARP_dev: Remove unstable interfaces and devices that should not have been in master
The `yarp::dev::IDepthSensor`, `yarp::dev::IRGBDSensor` interfaces and the
`RGBDSensorClient` and `RGBDSensorWrapper` devices were erroneously introduced
as a "stable", but they are still under development and therefore are not to
be used, and will be removed in the next tweak release (v2.3.65.1).
They will be probably re-introduced in the next stable release series
(v2.3.68).