YARP_OS: Refactor yarp::os::Contact class
The old interface causes a lot of unwanted copies, and it makes it hard
to understand what is happening to an object. The add*() methods create
a new copy of the object, and they were often used in this way:

  c = c.addName(name).addCarrier(carrier);

This, besides not being optimized makes it very complicated to
understand what happens in the code, and would be a lot more readable in
this way:

  c.setName(name);
  c.setCarrier(carrier);

Therefore instead of using add*() methods (now deprecated), it is now
recommended to use the corresponding set*() methods that will not
generate an extra Contact (unless, of course that's what you want, in
which case you should explicitly make a copy of the object, and then set
the required fields).

Instead of using the by*(), invalid() and empty() factory methods (now
deprecated), it is now recommended to use the constructors (several
constructors were added).

The byConfig() method was renamed fromConfig(), for coherence with
fromString().

The toURI() method now accepts an includeCarrier parameter (true by
default) to simplify the generation of the URI without the carrier, that
was often generated in this way, that is not intuitive and involves an
extra copy of the object:

  uri = c.addCarrier("").toURI();

It can now be simply generated in this way:

  uri = c.toURI(false);

The implementation details of the class are now hidden using the PIMPL
idiom.

Documentation was cleaned up and updated.