Fixed exception on thread join with error: `join failed: No such process`
The problem is that the join was called on a thread object in a process different from the one which created it.
This happened because of calls to `fork()`.

In particular we have to be careful when we call `fork()`: all the memory is duplicated but only one thread (the one calling `fork()`) survives.
In general it can be very dangerous to fork in a multithread application. It is only safe to call exec* as soon as possible after the fork. At that point the whole memory is substituted with the executed process.

To fix this issue I decided to completely remove the calls to `cleanBeforeExec` (which now is no more used, and probably should be removed).
In case exec* succeeds the function never returns. In case it fails, the process exit some statements after.