Yet another fix for static builds
In order to use the TARGET Libedit::libedit, since YARP is build
static, it is necessary that this target is known to the projects
depending on YARP. The easy fix is to use the path to the library
(hence ${Libedit_LIBRARIES}). At some point we should find a
proper fix for this.