Remove support for Atlas
The option to link Atlas library has been broken for a while, and
therefore it was removed.
On debian/ubuntu it is still possible to use the atlas library by
installing the libatlas-base-dev package, and use update-alternatives to
configure the alternatives for libblas.so, libblas.so.3, liblapack.so,
and liblapack.so.3.