OS/PortCore: do not interrupt input unit in closeMain
When an input port is deleted, all the ports connected to this port
should be disconnected first. Therefore a message is sent to these
ports, and these perform the actual disconnection.
The real disconnection is executed after sending a final message
on the socket.

The interrupt() call removed by this patch closes the socket
internally, therefore when the final message is sent by the output
port, the output socket is broken, hence the SIGPIPE.

Removing the interrupt here does not seem to have side effects, since
the socket is closed later, when the message arrives.