Partially revert "YARP_OS: YARP 2.1.2 is now a "long time ago""
This reverts commit bdb9540ab82669341c60e522f9aa77f06d4bab81.

As discussed on #921, it is not good to break compatibility between
master and devel branches, therefore for now only the "output" part is
changed and yarp devel will always "speak" the new protocol, but will
be able to understand both.

At some point after the next (few) release(s) we will break the
compatibility between devel and the old releases, but master and the
latest (few) release(s) will still be compatible.