OS/Contact: Fix C++11 issues
When YARP_OS is compiled with C++11 disabled, Contact move
constructor and move assignment operator are not compiled.
When the library is used, if C++11 is enabled, these are found in
the header file, and therefore they are used, but then they are not
found during linking.

Since Qt5 5.7 C++11 is required and it is automatically enabled, by
cmake, therefore if C++11 is disabled in YARP, these method should
be hidden.