CMake: Fix symbol visibility logic
Symbols in the public libraries that are not exported using YARP_xxx_API
are now always hidden.
This has been extensively tested on windows, but it was impossible to
enable hide them on linux and to enable the tests at the same time.

Symbols in the impl libraries that are exported using the
YARP_xxx_impl_API decorator instead are exported by default, but can be
hidden using the YARP_EXPERIMENTAL_FILTER_API cmake option (at the
moment this will fail to build).

Symbols in the impl libraries that are not exported using the
YARP_xxx_impl_API decorator are always hidden.