Changed how custom positions work in motorgui
Changed the ini syntax on how to specify custom positions.
Now the number of custom positions is no more hardcoded.
If no custom positions are defined in the ini file, no custom positions options are created in the Menu.
Instead, a submenu option is created and populated with the correct number of custom positions.

Added also menu option for the custom positons in the single part menubar.

Additionally, now the custom positions are named, so it is easier in the GUI to find and launch the correct custom positions when more than one is specified in the ini