Set and export variables before entering if-guard
yarp_configure_external_installation requires that certain variables
are defined on each call (e.g. the ones introduced by GNUInstallDirs).
Keep in mind that CMake macros and functions are visible in the global
scope once defined; therefore, one may have included this module in a
previous section of the CMake project tree and, as a result, fail to
load the entire file due to the presence of the if-guard. If this is
the case, and variables like CMAKE_INSTALL_DATADIR haven't been set in
the cache, yarp_configure_external_installation sees empty strings
instead and produces absolute paths.