Add BufferedConnectionWriter::applyConvertTextMode() const method
The applyConvertTextMode converts a text mode message to binary, if that
has been requested.

The const version of this method performs a const_cast, and calls the
non-const version. This makes possible calling it in const methods.
Conceptually this is not completely wrong because it does not modify
the external state of the class, but just some internal representation.