Don't add stray '_' char if bundle name is empty
Commit 6b02ae4 assumes one would wrap plugin declarations between
yarp_begin_plugin_library() and yarp_end_plugin_library(). However,
these make no sense in a DL world, that is, when all we want are
dynamically loaded libraries. In fact, those macros lead to the
creation of a dummy (i.e. 'empty', all it does is incrementing a
static counter) *static* library with .so extension. Apart from that,
a prefix is added to all CMake options associated with YARP plugins
these macros encompass. Such prefix was not mandatory, but said
commit was not prepared to handle an empty string for its value
(and this happens when the aforementioned macros were not used).

TL;DR: regression fix