Fix build with -fsanitize
Building thrift with clang sanitizer will cause errors when running
thrift, and it will make the build fail since it is executed during
the build.
Therefore the '-fsanitize' build flag is just removed in order to
avoid build failures, and allow to test the rest of YARP with clang
sanitizers.