idl/thrift: Do not use nested serialization for types annotated with "yarp.name"
This improves the compatibility with existing services returning Bottle
or other YARP types, and makes it possible to write a thrift file for
existing protocols.

On the other hand, this breaks compatibility with thrift services
returning Bottles or other annotated types, generated before this
commit. This is a breaking change, but should impact only a very
limited number of cases, and it can be easily fixed by regenerating the
files.

The unit tests was actually checking this behaviour, and it is therefore
updated according to this change.