yarpidl_thrift: Go back to have the helper object as a class member
The updateReply method of a portmonitor is not called on the same thread
that reads from the connection, therefore the return variable, as a
thread_local variable, causes it to be empty.

The flush performed in the WireWriter dtor, ensures that any
appendExternalBlock contained in helper serializationis actually written
before the helper is destroyed.

Fixes: #2021