os/idl/WireWriter: Add skip_tag flag to write methods.
This allows to write lists. For example, a list of string can be
written as:

  - BOTTLE_TAG_LIST
  - size
  - BOTTLE_TAG_STRING
  - lenght of the first element
  - first element
  - BOTTLE_TAG_STRING
  - lenght of the second element
  - second element
  - ...

but this is not optimized, and is supposed to be used for non-omogeneous
lists.
Alternatively, it can be optimized and written as:

  - BOTTLE_TAG_LIST | BOTTLE_TAG_STRING
  - size
  - lenght of the first element
  - first element
  - lenght of the second element
  - second element
  - ...

This syntax allows to write an array of numeric items, just by appending
the block of data after 'size'

Without the 'skip_tag' flag, 'writeString' would write an extra
'BOTTLE_TAG_STRING' for every element, and the parsing will fail on the
receiver.