Add documentation about local_rpc in frameTransformClient device
The ``frameTransformClient`` opens a utility RPC port named ``/ftClient/rpc`` independently from the ``ft_client_prefix``. This can raise a name conflict problem in case there is more than one client in the network.

The problem can be avoided by specifying the ``local_rpc`` option, but this was not documented. The commit adds the parameter in the documentation.